+++ src/gromo/containers/growing_container.py
@@ -1,3 +1,5 @@
+from typing import Any
+
 import torch
 
 from gromo.config.loader import load_config
@@ -31,11 +33,26 @@ class GrowingContainer(torch.nn.Module):
         """
         raise NotImplementedError
 
+    def set_scaling_factor(self, factor: float) -> None:
+        """Assign scaling factor to all growing layers
+
+        Parameters
+        ----------
+        factor : float
+            scaling factor
+        """
+        for layer in self._growing_layers:
+            if isinstance(layer, GrowingContainer):
+                layer.set_scaling_factor(factor)
+            elif isinstance(layer, GrowingModule):
+                layer.scaling_factor = factor
+                layer._scaling_factor_next_module.data[0] = factor
+
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """Forward pass through the network"""
         raise NotImplementedError
 
-    def extended_forward(self, x: torch.Tensor) -> torch.Tensor:
+    def extended_forward(self, x: torch.Tensor, mask: dict = {}) -> torch.Tensor:
         """Extended forward pass through the network"""
         raise NotImplementedError
 
@@ -86,17 +103,26 @@ class GrowingContainer(torch.nn.Module):
             if isinstance(layer, (GrowingModule, GrowingContainer)):
                 layer.compute_optimal_updates(*args, **kwargs)
 
-    def select_best_update(self) -> None:
+    def select_best_update(self) -> int:
         """Select the best update for growth procedure"""
         first_order_improvements: list[torch.Tensor] = [
             layer.first_order_improvement for layer in self._growing_layers
         ]
         best_layer_idx = torch.argmax(torch.stack(first_order_improvements))
-        self.currently_updated_layer_index = best_layer_idx
+        self.currently_updated_layer_index = best_layer_idx.item()
+        assert isinstance(self.currently_updated_layer_index, int), (
+            "Currently updated layer index is not an integer"
+            f" but {type(self.currently_updated_layer_index)}"
+        )
 
         for idx, layer in enumerate(self._growing_layers):
             if idx != best_layer_idx:
                 layer.delete_update()
+        return self.currently_updated_layer_index
+
+    def dummy_select_update(self, **_) -> int:
+        self.currently_updated_layer_index = 0
+        return 0
 
     def select_update(self, layer_index: int, verbose: bool = False) -> int:
         self.currently_updated_layer_index = layer_index
@@ -118,13 +144,13 @@ class GrowingContainer(torch.nn.Module):
         self,
     ) -> "GrowingModule | MergeGrowingModule | GrowingContainer":
         """Get the currently updated layer"""
-        assert self.currently_updated_layer_index is not None, "No layer to update"
+        assert isinstance(self.currently_updated_layer_index, int), "No layer to update"
         return self._growing_layers[self.currently_updated_layer_index]
 
-    def apply_change(self) -> None:
+    def apply_change(self, extension_size: int | None = None) -> None:
         """Apply changes to the model"""
         assert self.currently_updated_layer is not None, "No layer to update"
-        self.currently_updated_layer.apply_change()
+        self.currently_updated_layer.apply_change(extension_size=extension_size)
         self.currently_updated_layer.delete_update()
         self.currently_updated_layer_index = None
 
@@ -144,3 +170,14 @@ class GrowingContainer(torch.nn.Module):
         for layer in self._growing_layers:
             if isinstance(layer, (MergeGrowingModule, GrowingContainer)):
                 layer.update_size()
+
+    def weights_statistics(self) -> dict[str, Any]:
+        """Get the statistics of the weights in the growing layers.
+        Due to the recursive nature of the containers, the returned dictionary
+        contains nested dictionaries for each layer.
+        """
+        stats = {}
+        for module in self.modules():
+            if isinstance(module, GrowingModule):
+                stats[module.name] = module.weights_statistics()
+        return stats